+++ src/main/java/org/example/remedy/presentation/like/dto/response/LikeCountResponse.java
@@ -1,4 +1,4 @@
-package org.example.remedy.presentation.like.dto;
+package org.example.remedy.presentation.like.dto.response;
 
 public record LikeCountResponse(
         long likeCount
